Elegant solution for sharing "isLocalInstance" flag + move editor intro code to separate module
